+++ packages/shared/src/components/onboarding/index.ts
@@ -3,3 +3,4 @@ export * from './FilterOnboardingStep';
 export * from './FilterOnboardingV4';
 export * from './CreateFeedButton';
 export * from './FeedReadyMessage';
+export * from './OnboardingHeader';
